Eliminate compile warnings across local and CI matrix builds (#316)
* Start develop-wide compile warning cleanup


* Eliminate compile warnings across local and CI matrix builds

- Add version-portable Math::ThinSVD alias; replaces deprecated
  Eigen::BDCSVD constructor-options form at 6 call sites (AppleClang).
- Fix -Wsign-compare between Eigen signed size() and Index/size_t loop
  counters in PETSc assembly backends, GridFunction, MPI P1, and tests.
- Drop dead locals; mark assert-only reads [[maybe_unused]], keeping the
  assert(ierr == PETSC_SUCCESS) idiom.
- Silence vendored googletest -Wmaybe-uninitialized and mmg
  -Wstringop-overflow per-target (GCC only).
- Suppress Xcode 15 ld duplicate-libraries warning behind
  check_linker_flag.


* Style: clang-format touched lines, document ThinSVD constructor


* Use explicit version-guarded Eigen::BDCSVD instead of ThinSVD alias


* Fix AppleClang -Wall warnings surfaced by local test build

Release builds strip asserts, leaving assert-only reads unused: mark
found/ierr/vdim and the ghost-map structured binding [[maybe_unused]];
drop the dead D local in PETSc/IO/MFEM.h.


* Mark assert-only ierr [[maybe_unused]] in PETSc/Object.h


* Fix remaining gcc-14 warnings: P1 assert sign-compare, boost array-bounds

- Cast Eigen size() in the dofs/global assert in MPI P1.h.
- Suppress GCC's false-positive -Warray-bounds around the Boost.Container
  includes in Types.h (diagnostic is attributed to the Boost headers).


---------

+++ src/Rodin/Variational/GridFunction.h
@@ -1490,7 +1490,7 @@ namespace Rodin::Variational
       GridFunction& setData(const DataType& data, size_t offset = 0)
       {
         const auto sz = this->getFiniteElementSpace().getSize();
-        assert(offset + static_cast<size_t>(sz) <= data.size());
+        assert(offset + static_cast<size_t>(sz) <= static_cast<size_t>(data.size()));
         this->getData() = data.segment(offset, sz);
         return *this;
       }
